Add files via upload
I've finished. Although I couldn't properly add the video since I found it difficult to download I added it as a link. I added the images and finished the css'. If there's any problem regarding the images or the linking between the css and html id like to know, since working from the Mac has limitations with the linking g between images and stuff. The images in the file were originally for the adopt site but I had no more time and had to improvise.